Update AzuraCast Contact Information
Hello! AzuraCast recently launched a more powerful and better-documented homepage. Also, in the intervening time since this page was last updated, the project has moved to an independent GitHub organization, so the old URL needed updating in any case. I also clarified the description text to make it line up better with other projects described on the page.